[CALCITE-2444] Handle IN expressions when converting SqlNode to SQL (Zoltan Haindrich)
Previously RelToSqlConverter could not handle IN expressions because
it never saw them: SqlToRelConverter had converted them to OR, and if
an IN expression had been created internally - say via RelBuilder -
RelToSqlConverter would throw a ClassCastException.

Now we convert IN expressions to "IN" SQL, as you would expect.

Close apache/calcite#805