[CALCITE-2166] Cumulative cost of RelSubset.best RelNode is increased after calling RelSubset.propagateCostImprovements() for input RelNodes (Xiening Dai)
It's possible that Subset's best cost increases when input subset's best is
changed. In those cases, although input subset's cost is reduced, the row count
can increase which causes the increase of non-cumulative cost of parent rel. As
a result, the cost of parent rel can increase. If the parent rel happens to be
the best rel of a given subset, we currently do nothing. And this would lead to
the inconsistency of the rel node cost.

Fixing this by updating the best rel node cost if it's increased. Although this
approach won't garantee an optimal plan, at least it makes sure the memo is
consistent.

More details, please refer to JIRA -
https://issues.apache.org/jira/browse/CALCITE-2166

Close #1440