[CALCITE-3753] Introduce SubstitutionRule interface and execute substitutional rule first
A rule that implements SubstitutionRule interface indicates that the new
RelNode is definitely better than the old one, which will be pruned by setting
importance to 0. In addition, the rule matches for SubstitutionRule will be
inserted into the head of RuleQueue so that they can be executed as early as
possible.