[CALCITE-3889] Add apply(Mappings.Mapping) to RelTrait and RelTraitSet
RelTrait Collation, Distribution have key indices, when we pass down the
traitset to child or propagate to parent operator, we have to remap these keys.
It would be nice to have apply(Mappings.Mapping) on RelTrait and RelTraitSet.
RelDistribution already has the method, but we may want it on every RelTrait
except Convention.

Close #1912