[CALCITE-3924] Fix flakey test to handle TIMESTAMP and TIMESTAMP(0) correctly (neoReMinD)
The whole if block should be removed. Not only the DATETIME_TYPES has
this problem, all the types that allows precision could have such a
problem, the root cause is that BasicSqlType#toString and digest has different
handling of printing non specified precision.

One way to solve is to not fire caching of types without precision and
with default precision.

close apache/calcite#1916