[CALCITE-3927] RelSubset is not fired for rule when set gets merged (Botong Huang)
In VolcanoPlanner, when set gets merged, planner fires rules again for RelNodes
in both sets, but not for RelSubset. We might miss something because of this.

If all the logical transformation rules and physical implementation rules are
separated out in different stage and physical rules don't do logical work, we
might be OK. But the reality is that all the things are mixed together at the
moment.

Close #1922