[CALCITE] Add support for colon separator in compound identifiers (#124)
This change also modifies the way JSON names are extracted because the previous logic erroneously used the Expression() parsing function to extract only a string literal or an identifier. This did not work once <COLON> was added as an allowed part of a compound identifier for an input key/value pair. For example, "foo: bar".

PiperOrigin-RevId: 319158260