[CALCITE] Moved parser test files from core/server/babel into parsing directory (#141)
A lot of the files in the core/server/babel modules say they were deleted, but the files in the parsing directory were actually deleted and the module test files were moved to the parsing directory. The goal was to keep the git blame history since copying the files overwrites the original history while moving the files keeps it.