[CALCITE] Validation: Allow unit tests to specify parser (#306)
There were a 11 tests in SqlValidatorTest that failed when executed using the Dialect1 parser. I moved most of these to DefaultDialectValidatorTest with the exception of 1 that was MySQL specific which I moved to MySQLValidatorTest.

The @LocaleEnUs annotation was removed because it didn't seem to serve much of a purpose and was throwing an initilizationError when I tried to extend SqlValidatorTest.