[CALCITE] Unknown compound identifiers (#323)
When tables are specified via compound identifiers like foo.bar, this will now result in prefixes getting added as subschemas to the root schema, and tables getting added to the appropriate subschemas (ex. "foo" will be added as a subschema to the root, and "bar" will be added as a table to subschema "foo").